+++ lib/search_flip/index.rb
@@ -254,7 +254,7 @@ module SearchFlip
         :page, :per, :search, :highlight, :suggest, :custom, :find_in_batches, :find_results_in_batches,
         :find_each, :find_each_result, :failsafe, :total_entries, :total_count, :timeout, :terminate_after,
         :records, :results, :must, :must_not, :should, :preference, :search_type, :routing,
-        :track_total_hits, :explain
+        :track_total_hits, :explain, :http_timeout
 
       # Override to specify the type name used within Elasticsearch. Recap,
       # this gem uses an individual index for each index class, because