+++ spec/search_flip/index_spec.rb
@@ -14,7 +14,7 @@ RSpec.describe SearchFlip::Index do
       :total_entries, :total_count, :terminate_after, :timeout, :records, :results,
       :must, :must_not, :should, :find_each_result,
       :find_results_in_batches, :preference, :search_type, :routing,
-      :track_total_hits, :explain
+      :track_total_hits, :explain, :http_timeout
     ]
 
     methods.each do |method|